Move test fixture for hi_l1a into conftest.py for reuse
